Moved sampling-based tests out of unit and into an exectuable target.
